refactor[ia7.2]: renamed service to communicator and changed folder
