+++ HealthcareSystem/IntegrationAdapters/Controllers/SftpController.cs
@@ -1,6 +1,6 @@
-﻿using Backend.Model.Pharmacies;
+﻿using Backend.Communication.SftpCommunicator;
+using Backend.Model.Pharmacies;
 using Backend.Service.DrugConsumptionService;
-using Backend.Service.SftpService;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 
@@ -8,12 +8,12 @@ namespace IntegrationAdapters.Controllers
 {
     public class SftpController : Microsoft.AspNetCore.Mvc.Controller
     {
-        private readonly ISftpService _sftpService;
+        private readonly ISftpCommunicator _sftpCommunicator;
         private readonly IDrugConsumptionService _drugConsumptionService;
 
-        public SftpController(ISftpService sftpService, IDrugConsumptionService drugConsumptionService)
+        public SftpController(ISftpCommunicator sftpCommunicator, IDrugConsumptionService drugConsumptionService)
         {
-            _sftpService = sftpService;
+            _sftpCommunicator = sftpCommunicator;
             _drugConsumptionService = drugConsumptionService;
         }
 
@@ -31,7 +31,7 @@ namespace IntegrationAdapters.Controllers
             var reportFileName = $"{dateRange.From:yyyy-MM-dd}-to-{dateRange.To:yyyy-MM-dd}-report.json";
 
             System.IO.File.WriteAllText(reportFileName, json);
-            _sftpService.UploadFile(reportFileName, "/PSW-uploads/" + reportFileName);
+            _sftpCommunicator.UploadFile(reportFileName, "/PSW-uploads/" + reportFileName);
 
             TempData["Success"] = "Report successfully created and uploaded!";
             return RedirectToAction("Index");
